git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@16995 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/statistic.c
@@ -2395,7 +2395,7 @@ MagickExport ChannelStatistics *GetImageChannelStatistics(const Image *image,
     channel_statistics[i].standard_deviation=sqrt(
       channel_statistics[i].variance-(mean*mean));
   }
-  for (i=0; i < (MaxMap+1U); i++)
+  for (i=0; i < (ssize_t) (MaxMap+1U); i++)
   {
     histogram[i].red/=area;
     channel_statistics[RedChannel].entropy+=-histogram[i].red*
